Move 'FindStageByTwoPartKey'.
